requirements.txt
playwright
selenium
requests
pillow
imagehash
pyodbc
python-dotenv
sv_ttk
openpyxl
